==> 05_rolle_kbob.Rmd <==
---
title: "05 - KBOB - Ökobilanz-Kennwerte"
author: "Curdin Derungs"
date: "August 2018"
output:
  html_document:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
---

```{r setup, include=FALSE}

#loading libraries
library(sp)
library(leaflet)
library(ggplot2)
library(gridExtra)
library(data.table)
library(dplyr)

#empty workspace
rm(list=ls())

#load helpers
source("999_helpers.R")

#encoding
Sys.setlocale(category = "LC_ALL", locale = "en_US.UTF-8")
```

In diesem Skript Nutzungsgrade und KBOB Werte entsprechend dem Energieträger den einzelnen Gebäuden zugewiesen.

#Daten einlesen
Die Daten aus dem vorhergehenden Arbeitsschritt werden eingelesen.
```{r reader av}
load("output/04_output/geom/04_rolle_bld_out.Rdata")

ng<-as.data.frame(fread(paste(dataPath,"16_kbob/nutzungsgrad.csv",sep=""),encoding="Latin-1"))
kbob<-as.data.frame(fread(paste(dataPath,"16_kbob/kbob.csv",sep=""),encoding="Latin-1"))
bestand<-as.data.frame(fread(paste(dataPath,"16_kbob/SIA2024_2015_Bestand.csv",sep=""),encoding="Latin-1"))
```

#Verteilung Energieträger
```{r energietraeger, fig.height=10, fig.width=6, warning=FALSE, paged.print=FALSE}
p1<-ggplot(rolle_bld_out@data, aes(x=GENHZ_text))+
  geom_bar()+
  theme_minimal()+
  ggtitle(paste("Energie (Anzahl Gebaeude)",sep=""))+
  theme(axis.text.x = element_text(angle = 45, hjust = 1,size=14))+
  xlab("")

p2<-ggplot(rolle_bld_out@data, aes(x=GENHZ_text,y=GAREA))+
  geom_bar(stat = "identity")+
  theme_minimal()+
  ggtitle(paste("Energie (Flaechensummen Gebaeude)",sep=""))+
  theme(axis.text.x = element_text(angle = 45, hjust = 1,size=14))+
  xlab("")

p3<-ggplot(rolle_bld_out@data, aes(x=GENHZ_text,y=ebf))+
  geom_bar(stat = "identity")+
  theme_minimal()+
  ggtitle(paste("Energie (EBF-Summe Gebaeude)",sep=""))+
  theme(axis.text.x = element_text(angle = 45, hjust = 1,size=14))+
  xlab("")

p.all<-grid.arrange(p1,p2,p3,ncol=1)
p.all
```
  
Gas und Heizöl domninieren als Energieträger in Rolle und MS-Rolle. Für eine nicht zu vernachlässigbare Anzahl Gebäude ist der Energieträger nicht bekannt (NA). Es handelt sich wahrscheinlich um nicht beheizte Lagerhallen (müsste überprüft werden!).

#Nutzungsgrad

Nutzungsgrade um von der Endenergie, wie beispielsweise durch die AWEL-Methode geschätzt, auf die Nutzenergie zu schliessen werden entsprechend der SIA 380.

```{r nutzungsgrad, warning=FALSE, paged.print=FALSE}
rolle_dat<-rolle_bld_out@data
rolle_dat<-left_join(rolle_dat, ng, by = c('GENHZ_text'='energietraeger'))
```

#KBOB Faktoren

Die KBOB Faktoren stammen von folgendem Link [Link](https://www.kbob.admin.ch/dam/kbob/de/dokumente/Publikationen/Nachhaltiges%20Bauen/Archiv_2005-2009/Liste%20Oekobilanzdaten%20im%20Baubereich%202009-1-2016-gerundet-kWh.zip.download.zip/Liste%20Oekobilanzdaten%20im%20Baubereich%202009-1-2016-gerundet-kWh.zip). 

```{r kbob, warning=FALSE, paged.print=FALSE}
rolle_dat<-left_join(rolle_dat, kbob, by = c('GENWW_text'='energietraeger'))


rolle_bld_out@data<-rolle_dat
```

#Energiebedarf unterschiedlicher Verbraucher

```{r bestand, warning=FALSE, paged.print=FALSE}
bestand$siaLabel<-NULL
names(bestand)<-paste("bestand_",names(bestand),sep="")
rolle_dat<-left_join(rolle_dat, bestand, by = c('siaNr'='bestand_siaNr'))

rolle_bld_out@data<-rolle_dat
```

##Save all Information
```{r save}
#buildings speichern
save(rolle_bld_out,file="output/05_output/geom/05_rolle_bld_out.Rdata")
```
